# Refactor labels package to reduce cyclomatic complexity