--- a/resolver/resolver_test.go
+++ b/resolver/resolver_test.go
@@ -94,7 +94,7 @@ func fakeDNS(port int) (*Resolver, error) {
 	}
 
 	masters := []string{"144.76.157.37:5050"}
-	spec := labels.ForRFC952()
+	spec := labels.RFC952
 	res.rs = &records.RecordGenerator{}
 	res.rs.InsertState(sj, "mesos", "mesos-dns.mesos.", "127.0.0.1", masters, spec)