This change set reduces the cyclomatic complexity of the labels package and reduces its API surface.
With this patch, we're trading label generation performance for maintainability and easier correctness under the premise that the 3x slow-down due to allocations and garbage collection is acceptable even at large scales. On my Macbook, 1M tasks take 1.59748894s to generate without any code optimisation done yet. Here are the more detailed benchmarks and profiling information:

## Benchmarks
```
benchmark                       old ns/op     new ns/op     delta
BenchmarkDomainFrag_RFC952      1192          4198          +252.18%
BenchmarkDomainFrag_RFC1123     1206          3937          +226.45%

benchmark                       old allocs    new allocs    delta
BenchmarkDomainFrag_RFC952      13            48            +269.23%
BenchmarkDomainFrag_RFC1123     14            49            +250.00%

benchmark                       old bytes     new bytes     delta
BenchmarkDomainFrag_RFC952      192           984           +412.50%
BenchmarkDomainFrag_RFC1123     192           992           +416.67%
```

## CPU profiles
### Previous implementation
![cpu](https://cloud.githubusercontent.com/assets/67471/9003642/e3f4d776-376f-11e5-9bff-880ca21a7274.png)

### New implementation
![cpu](https://cloud.githubusercontent.com/assets/67471/9003664/118fa26a-3770-11e5-93d2-ddd5d5608507.png)

## Other ideas
1) Why don't we cache the results of label generation? These should never change for the same input.
2) Why is the whole state rebuilt on every state load? It'd be more efficient to incrementally apply state changes.