@kozyraki 2015-07-31
Hi Tomas, 

first of all, there are really cool graphs and major kudos for measuring the performance impact of your changes. Great engineering discipline my fried, congrats!

Can you explain a little what where does the improvement in maintainability and correctness comes from? The graph in the bottom looks more complex than the one in the top. Where do the benefits come from? 

About the 3x slowdown in generation: it is probably ok. You are not changing the load on the master, which would be really bad with the current implementation of Mesos-DNS for large systems. 1.6s of generation for 1M tasks every 60s is not bad. We may want to double-check what happens to serving during generation. If Mesos-DNS has a couple of cores in most deployments, the impact of longer record generation on record serving should not be that bad. But we should double-check it. If this works out and the maintainability advantages are major, it is ok with me. 

Now, about your other questions:
- The reason we don’t cache the results of record generation is the following. Our only way of interacting with the master is to poll for the whole task state. If we wanted to cache records, we would have to scan through the new task state each time, find the task that are the same and skip them (most of them), find the tasks that changed and update them, and then find the missing tasks and erase their records. It’s actually simpler to regenerate everything from scratch and get it over with. I also bet it is faster as the diff would have non trivial overheads. 
- Ideally, we should not get the whole task state every 60 sec. We should just get task updates on launch or finish/failure and that would allow us to immediately generate/remove records accordingly. Unfortunately, Mesos does not have such an interface right now. There are many discussions on how to do this. Once there is one of them, we will switch Mesos-DNS. It will solve many issues around scaling. 

Christos

> On Jul 31, 2015, at 10:40 AM, Tomás Senart <notifications@github.com> wrote:
> 
> This change set reduces the cyclomatic complexity of the labels package and reduces its API surface.
> With this patch, we're trading label generation performance for maintainability and easier correctness under the premise that the 3x slow-down due to allocations and garbage collection is acceptable even at large scales. On my Macbook, 1M tasks take 1.59748894s to generate without any code optimisation done yet. Here are the more detailed benchmarks and profiling information:
> 
> benchmark                       old ns/op     new ns/op     delta
> BenchmarkDomainFrag_RFC952      1192          4198          +252.18%
> BenchmarkDomainFrag_RFC1123     1206          3937          +226.45%
> 
> benchmark                       old allocs    new allocs    delta
> BenchmarkDomainFrag_RFC952      13            48            +269.23%
> BenchmarkDomainFrag_RFC1123     14            49            +250.00%
> 
> benchmark                       old bytes     new bytes     delta
> BenchmarkDomainFrag_RFC952      192           984           +412.50%
> BenchmarkDomainFrag_RFC1123     192           992           +416.67%
> CPU profiles
> 
> Previous implementation
> 
>  <https://cloud.githubusercontent.com/assets/67471/9003642/e3f4d776-376f-11e5-9bff-880ca21a7274.png>
> New implementation
> 
>  <https://cloud.githubusercontent.com/assets/67471/9003664/118fa26a-3770-11e5-93d2-ddd5d5608507.png>
> Other ideas
> 
> 1) Why don't we cache the results of label generation? These should never change for the same input.
> 2) Why is the whole state rebuilt on every state load? It'd be more efficient to incrementally apply state changes.
> 
> You can view, comment on, or merge this pull request online at:
> 
>   https://github.com/mesosphere/mesos-dns/pull/207 <https://github.com/mesosphere/mesos-dns/pull/207>
> Commit Summary
> 
> labels: Refactor and reduce complexity
> Adjust cyclomatic complexity threshold
> File Changes
> 
> M circle.yml <https://github.com/mesosphere/mesos-dns/pull/207/files#diff-0> (2)
> M records/generator.go <https://github.com/mesosphere/mesos-dns/pull/207/files#diff-1> (13)
> M records/generator_test.go <https://github.com/mesosphere/mesos-dns/pull/207/files#diff-2> (3)
> D records/labels/domfrag.go <https://github.com/mesosphere/mesos-dns/pull/207/files#diff-3> (41)
> D records/labels/domfrag_test.go <https://github.com/mesosphere/mesos-dns/pull/207/files#diff-4> (79)
> A records/labels/labels.go <https://github.com/mesosphere/mesos-dns/pull/207/files#diff-5> (83)
> A records/labels/labels_test.go <https://github.com/mesosphere/mesos-dns/pull/207/files#diff-6> (179)
> D records/labels/spec.go <https://github.com/mesosphere/mesos-dns/pull/207/files#diff-7> (49)
> D records/labels/tables.go <https://github.com/mesosphere/mesos-dns/pull/207/files#diff-8> (97)
> D records/labels/tables_test.go <https://github.com/mesosphere/mesos-dns/pull/207/files#diff-9> (103)
> M resolver/resolver_test.go <https://github.com/mesosphere/mesos-dns/pull/207/files#diff-10> (2)
> Patch Links:
> 
> https://github.com/mesosphere/mesos-dns/pull/207.patch <https://github.com/mesosphere/mesos-dns/pull/207.patch>
> https://github.com/mesosphere/mesos-dns/pull/207.diff <https://github.com/mesosphere/mesos-dns/pull/207.diff>
> —
> Reply to this email directly or view it on GitHub <https://github.com/mesosphere/mesos-dns/pull/207>.
>