@sttts 2015-07-31
Moving the optimizations to another PR is fine, especially because the performance is still "good enough".